Split vscode-dependent code
