+++ vscode/src/services/treeViewItems.ts
@@ -2,7 +2,8 @@ import { FeatureFlag } from '@sourcegraph/cody-shared/src/experimentation/Featur
 
 import { getChatPanelTitle } from '../chat/chat-view/chat-helpers'
 import { CODY_DOC_URL, CODY_FEEDBACK_URL, DISCORD_URL } from '../chat/protocol'
-import { releaseNotesURL, releaseType, version } from '../version'
+import { releaseNotesURL, releaseType } from '../release'
+import { version } from '../version'
 
 import { localStorage } from './LocalStorageProvider'
